+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/security/SecurityConfig.java
@@ -115,6 +115,7 @@ public class SecurityConfig {
                         "/data-exchange-request/**",
                         "/data-exchange-approval/**",
                         "/demand-and-capacity-notification/**",
+                        "/aggregated-data/**",
                         "/planned-production/**",
                         "/material-demand/**",
                         "/delivery-information/**",